Face_Recognizers_package/requirements.txt
opencv-python
scikit-image
onnxruntime-gpu
imageio
scipy
numpy
pyyaml